Adding completed MeshRenderer component that can be added to any GameObject, converted Vector3 and Vector2 to use floats, UniformBuffer translation updates based on Transform component so PlayerController can move the MeshRenderer of a GameObject it is attached to

+++ Dilgine/Dilgine/Scripts/Core/Vector2.h
@@ -6,11 +6,11 @@
 struct Vector2
 {
 public:
-	int x, y;
+	float x, y;
 
 	Vector2() { x = 0; y = 0; }
-	Vector2(int val) { x = val; y = val; }
-	Vector2(int valX, int valY) { x = valX; y = valY; }
+	Vector2(float val) { x = val; y = val; }
+	Vector2(float valX, float valY) { x = valX; y = valY; }
 
 	float Magnitude() { return sqrt((x * x) + (y * y)); }
 	Vector2 Normalized() {
@@ -22,10 +22,10 @@ public:
 
 	Vector2 operator+(const Vector2& rhs) { return Vector2(x + rhs.x, y + rhs.y); }
 	Vector2 operator-(const Vector2& rhs) { return Vector2(x - rhs.x, y - rhs.y); }
-	Vector2 operator*(const int rhs) { return Vector2(x * rhs, y * rhs); }
+	Vector2 operator*(const float rhs) { return Vector2(x * rhs, y * rhs); }
 	void operator+=(const Vector2& rhs) { x += rhs.x; y += rhs.y; }
 	void operator-=(const Vector2& rhs) { x -= rhs.x; y -= rhs.y; }
-	void operator*=(const int rhs) { x *= rhs; y *= rhs; }
+	void operator*=(const float rhs) { x *= rhs; y *= rhs; }
 
 	bool operator==(const Vector2& rhs) const { return (x == rhs.x && y == rhs.y); }
 	bool operator!=(const Vector2& rhs) const { return (x != rhs.x || y != rhs.y); }
